优化profile ui

+++ V2rayU/Views/Profile/ProfileListView.swift
@@ -14,11 +14,13 @@ struct ProfileListView: View {
     @State private var selection: Set<ProfileModel.ID> = []
     @State private var selectedRow: ProfileModel? = nil
     @State private var pingRow: ProfileModel? = nil
+    @State private var shareRow: ProfileModel? = nil
     @State private var selectGroup: String = ""
     @State private var searchText = ""
     @State private var draggedRow: ProfileModel?
     @State private var selectAll: Bool = false
     @State private var showPingSheet: Bool = false
+    @State private var showShareSheet: Bool = false
 
     var filteredAndSortedItems: [ProfileModel] {
         let filtered = viewModel.list.filter { item in
@@ -82,11 +84,13 @@ struct ProfileListView: View {
                     }
                     .buttonStyle(.bordered)
                     Button(action: {
-                        withAnimation {
-                            for selectedID in self.selection {
-                                viewModel.delete(uuid: selectedID)
+                        if showConfirmAlertSync(title: "Are you sure you want to delete the selected Proxy Profiles?", message: "This action cannot be undone.") {
+                            withAnimation {
+                                for selectedID in self.selection {
+                                    viewModel.delete(uuid: selectedID)
+                                }
+                                selection.removeAll()
                             }
-                            selection.removeAll()
                         }
                     }) {
                         Label("删除", systemImage: "trash")
@@ -101,10 +105,10 @@ struct ProfileListView: View {
                     .disabled(viewModel.list.isEmpty)
                     // 分享
                     Button(action: {
-                        
                     }) {
                         Label("分享", systemImage: "square.and.arrow.up")
-                    }
+                    }.disabled(selection.isEmpty)
+                        .buttonStyle(.bordered)
                 }.padding(.horizontal, 10)
                 // 表格主体
                 Table(of: ProfileModel.self, selection: $selection, sortOrder: $sortOrder) {
@@ -116,7 +120,7 @@ struct ProfileListView: View {
                     }
                     .width(30)
                     TableColumn("Type") { row in
-                        Text(row.`protocol` == .shadowsocks ? "ss" : row.`protocol`.rawValue)
+                        Text(row.protocol == .shadowsocks ? "ss" : row.protocol.rawValue)
                             .font(.system(size: 13))
                             .onTapGesture(count: 2) { selectedRow = row }
                     }
@@ -162,7 +166,6 @@ struct ProfileListView: View {
                     }
                     .dropDestination(for: ProfileModel.self, action: handleDrop)
                 }
-                
             }
             .background(.ultraThinMaterial)
             .border(Color.gray.opacity(0.1), width: 1)
@@ -174,7 +177,7 @@ struct ProfileListView: View {
                 loadData()
             }
         }
-        .sheet(item: $pingRow) { row in
+        .sheet(item: $pingRow) { _ in
             ProfilePingView(profile: pingRow, isAll: false) {
                 pingRow = nil
             }
@@ -183,58 +186,165 @@ struct ProfileListView: View {
             ProfilePingView(profile: nil, isAll: true) {
                 showPingSheet = false
             }
-        }.task { loadData() }
+        }
+        .sheet(item: $shareRow) { _ in
+            ProfileShareView(profile: shareRow, isAll: false) {
+                shareRow = nil
+            }
+        }
+        .sheet(isPresented: $showShareSheet) {
+            ProfileShareView(profile: nil, isAll: true) {
+                showShareSheet = false
+            }
+        }
+        .task { loadData() }
     }
 
     // 处理拖拽排序逻辑:
     // 参考: https://levelup.gitconnected.com/swiftui-enable-drag-and-drop-for-table-rows-with-custom-transferable-aa0e6eb9f5ce
     func handleDrop(index: Int, rows: [ProfileModel]) {
-        guard let firstRow = rows.first, let firstRemoveIndex = list.firstIndex(where: { $0.id == firstRow.id }) else { return }
+        guard let firstRow = rows.first, let firstRemoveIndex = viewModel.list.firstIndex(where: { $0.uuid == firstRow.uuid }) else { return }
 
-        list.removeAll(where: { row in
-            rows.contains(where: { insertRow in insertRow.id == row.id })
+        viewModel.list.removeAll(where: { row in
+            rows.contains(where: { insertRow in insertRow.uuid == row.uuid })
         })
 
-        list.insert(contentsOf: rows, at: index > firstRemoveIndex ? (index - 1) : index)
+        viewModel.list.insert(contentsOf: rows, at: index > firstRemoveIndex ? (index - 1) : index)
+        print("handleDrop: \(index) \(rows.count)")
+        // 更新排序
+        viewModel.updateSortOrderInDBAsync()
     }
 
     private func contextMenuProvider(item: ProfileModel) -> some View {
         Group {
+            Button {
+                chooseItem(item: item)
+            } label: {
+                Label("Choose", systemImage: "checkmark.circle")
+            }
+
+            Button {
+                self.pingRow = item
+            } label: {
+                Label("Test Latency", systemImage: "speedometer")
+            }
+
             Divider()
-            Button("Choose") {
+
+            Button {
+                copyItem(item: item)
+            } label: {
+                Label("Copy URI", systemImage: "doc.on.doc")
             }
-            Button("Duplicate") {
+
+            Button {
+                self.shareRow = item
+            } label: {
+                Label("QRCode", systemImage: "qrcode")
             }
-            Button("Copy") {
-            }
-            Button("Ping") {
-                self.pingRow = item
-            }
-            Button("Share") {
-                self.pingRow = item
-            }
+
             Divider()
-            Button("🔝 Move to Top") {
-                self.pingRow = item
+
+            Button(action: {
+                moveToTop(item: item)
+            }) {
+                Label("Move to Top", systemImage: "arrow.up.to.line")
             }
-            Button("Move to Bottom") {
-                self.pingRow = item
+
+            Button(action: {
+                moveToBottom(item: item)
+            }) {
+                Label("Move to Bottom", systemImage: "arrow.down.to.line")
             }
-            Button("↑ Move Up") {
-                self.pingRow = item
+
+            Button(action: {
+                moveUp(item: item)
+            }) {
+                Label("Move Up", systemImage: "chevron.up")
             }
-            Button("↓ Move Down") {
+
+            Button(action: {
+                moveDown(item: item)
+            }) {
+                Label("Move Down", systemImage: "chevron.down")
             }
+
             Divider()
-            Button("Edit") {
+
+            Button {
+                duplicateItem(item: item)
+            } label: {
+                Label("Duplicate", systemImage: "plus.square.on.square")
+            }
+
+            Button {
                 self.selectedRow = item
+            } label: {
+                Label("Edit", systemImage: "pencil")
             }
-            Button("Delete") {
-                // Handle another action
-                print("item.uuid", item.id, item.uuid)
-                viewModel.delete(uuid: item.uuid)
+
+            Button {
+                if showConfirmAlertSync(title: "Are you sure you want to delete this Proxy Profile?", message: "This action cannot be undone.") {
+                    viewModel.delete(uuid: item.uuid)
+                }
+            } label: {
+                Label("Delete", systemImage: "trash")
+                    .foregroundColor(.red)
             }
         }
+    }
+
+    private func chooseItem(item: ProfileModel) {
+        // 选择当前配置
+        AppState.shared.setRunning(uuid: item.uuid)
+    }
+
+    private func duplicateItem(item: ProfileModel) {
+        guard let index = viewModel.list.firstIndex(where: { $0.id == item.id }) else { return }
+        let newItem = item.clone()
+        newItem.index = index + 1 // 设置新的索引
+        viewModel.upsert(item: newItem)
+        viewModel.updateSortOrderInDBAsync()
+    }
+
+    private func copyItem(item: ProfileModel) {
+        // 复制到剪贴板
+        let pasteboard = NSPasteboard.general
+        pasteboard.clearContents()
+        let profileString = ShareUri.generateShareUri(item: item)
+        if pasteboard.setString(profileString, forType: .string) {
+            print("Copied to clipboard: \(profileString)")
+            alertDialog(title: "Copied", message: "")
+        } else {
+            print("Failed to copy to clipboard")
+            alertDialog(title: "Failed to copy to clipboard", message: "")
+        }
+    }
+
+    private func moveToTop(item: ProfileModel) {
+        guard let index = viewModel.list.firstIndex(where: { $0.id == item.id }) else { return }
+        viewModel.list.remove(at: index)
+        viewModel.list.insert(item, at: 0)
+        viewModel.updateSortOrderInDBAsync()
+    }
+
+    private func moveToBottom(item: ProfileModel) {
+        guard let index = viewModel.list.firstIndex(where: { $0.id == item.id }) else { return }
+        viewModel.list.remove(at: index)
+        viewModel.list.append(item)
+        viewModel.updateSortOrderInDBAsync()
+    }
+
+    private func moveUp(item: ProfileModel) {
+        guard let index = viewModel.list.firstIndex(where: { $0.id == item.id }), index > 0 else { return }
+        viewModel.list.swapAt(index, index - 1)
+        viewModel.updateSortOrderInDBAsync()
+    }
+
+    private func moveDown(item: ProfileModel) {
+        guard let index = viewModel.list.firstIndex(where: { $0.id == item.id }), index < viewModel.list.count - 1 else { return }
+        viewModel.list.swapAt(index, index + 1)
+        viewModel.updateSortOrderInDBAsync()
     }
 
     private func loadData() {
